+++ coordinating/inmemory/optimistic_update.go
@@ -1,7 +1,7 @@
 package inmemory
 
 import (
-	"github.com/paust-team/shapleq/agent/logger"
+	"github.com/paust-team/shapleq/logger"
 	"github.com/paust-team/shapleq/qerror"
 	"sync"
 )
